준석_리뷰 디테일 출력 완료 / 상세리뷰 영어로 쓰는 경우 입력창 벗어남 > 이어서 수정할 것.

+++ Mealchelin/src/main/java/com/mealchelin/mvc/review/model/vo/Review.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class Review {
-	    private Long reviewNo; // 리뷰 번호
+	    private int reviewNo; // 리뷰 번호
 	    
 	    private Double rated; // 별점
